updated installation steps for a working setup
CoppeliaSim updated their repositories and the installation setup was kinda broken. CoppeliaSim 4.0.0 has a broken python package that needs to be updated from another repository. The ROS Plugin for CoppeliaSim was also broken for 4.0.0 and for Ubuntu 16. This was now fixed using specific versions of the packages and a new install procedure.